🤡(backend) mock jwks endpoint
This introduce a new endpoint, which is a read-only URL that returns JWKSs
as JSON objects in response to GET requests.

JWKSs are sets of JWKs, which are JavaScript Object Notation (JSON) data structures
that represent cryptographic keys.

This /jwks route might be refactored to the be located at a well-known location.