✨(backend) expose cryptographic key for encryption
This commit focuses on generating a valid keyset for the project. The following
points need to be discussed:

- Error Handling
- Settings' Names
- Storing Private Key as a decoded string or a PEM file

The JWK structure and parameters align with the implementation of France
Connect (FC) mocked data provider, named resource server in Oauth spec.

The current implementation is fully parameterizable, allowing for easy
configuration changes via Django settings. For instance, switching to another
key type is as simple as updating the settings.

However, one critique of this design is its lack of extensibility. If the
decision is made to offer more than one encryption method, this view will
require refactoring.

This commit is laying the foundation for further discussions and improvements.