♻️(backend) encapsulate cryptographic key loading
I explained the responsibilities of the private key in the context of the
resource server.

Consider enhancing this utility function, as it currently imposes an
opinionated approach to loading the private key and passing parameters to the
joserfc utilities.

The documentation could be more concise to avoid excessive detail.

The current API and return type are tightly coupled with joserfc usage and
types. Should we consider exposing a generic Python object to simplify the
interface?