✨(backend) add get_settings method
Copied from mozzilla-django-oidc repository, add a static method
on my authentication backend, which get settings from Django settings,
or give a default value if provided as an arg.

It will be use a lot when accessing Django settings in the next methods.
Might be discussed, and considered as an unnecessary wrapper.