🚧(backend) refactor token introspection to separate responsibilities
Note: This commit is work in progress and has not been tested on my
machine. I'm pushing it now to gather early feedback.

Based on @Jonathan Perret's recommendations, I've split the responsibilities
of resource server authentication into three distinct classes, each mapping
to specific roles:

- A DRF authentication class
- A client to interact with the Authorization Server
- A client to interact with the Resource Server

In the OAuth 2.0 context, the Authorization Server is equivalent to the OIDC
provider. This separation aims to improve the readability of the code and
facilitate unit testing.

Additionally, this refactor provides better documentation of the technical
specifications followed by my code.