✨(frontend) add search component
One of the main aspects of La Régie is that there is an entity search in all
modules of the application.It is possible to search for a group, search for a
user to add them to a group with rights, or simply contacts.
We are therefore adding this component which will be reusable and adaptable to
several needs.